chore: rename input_trajectory to trajectory
