+++ src/trajectory.rs
